refactor(sinks): extract backpressure handling into reusable component
NetworkSink carried inline buffering + hysteresis logic for handling
a full outbound channel. Lift it into BackpressureHandler so other
sinks (MQTT, future plugins) can reuse the same protocol:

  - onFull(buffer, controller): stash buffer; acquire backpressure once
    the deque reaches upperThreshold. One buffer is kept "pending" so
    the sink always has something to retry, preventing a deadlock
    where both deque and pending slot are empty while upstream is
    blocked.
  - onSuccess(controller): clear pending; release backpressure once
    the deque drops to lowerThreshold; return the next buffered tuple
    so the sink can immediately retry.

State is held in folly::Synchronized. No behavior change for
NetworkSink — it now delegates to BackpressureHandler.
