+++ nes-sinks/include/Sinks/NetworkSink.hpp
@@ -17,16 +17,14 @@
 #include <atomic>
 #include <chrono>
 #include <cstddef>
-#include <deque>
 #include <optional>
 #include <ostream>
 #include <string>
 #include <unordered_map>
 #include <vector>
 #include <Configurations/Descriptor.hpp>
-#include <Identifiers/Identifiers.hpp>
-#include <Identifiers/NESStrongType.hpp>
 #include <Runtime/TupleBuffer.hpp>
+#include <Sinks/BackpressureHandler.hpp>
 #include <Sinks/Sink.hpp>
 #include <Sinks/SinkDescriptor.hpp>
 #include <folly/Synchronized.h>
@@ -37,29 +35,6 @@
 
 namespace NES
 {
-class BackpressureHandler
-{
-    struct State
-    {
-        bool hasBackpressure = false;
-        std::deque<TupleBuffer> buffered;
-        SequenceNumber pendingSequenceNumber = INVALID<SequenceNumber>;
-        ChunkNumber pendingChunkNumber = INVALID<ChunkNumber>;
-    };
-
-    folly::Synchronized<State> stateLock;
-    size_t upperThreshold;
-    size_t lowerThreshold;
-
-public:
-    /// @param upperThreshold Number of buffered tuples at which backpressure is acquired.
-    /// @param lowerThreshold Number of buffered tuples at which backpressure is released.
-    explicit BackpressureHandler(size_t upperThreshold = 1, size_t lowerThreshold = 0); /// NOLINT(fuchsia-default-arguments-declarations)
-    std::optional<TupleBuffer> onFull(TupleBuffer buffer, BackpressureController& backpressureController);
-    std::optional<TupleBuffer> onSuccess(BackpressureController& backpressureController);
-    bool empty() const;
-};
-
 class NetworkSink final : public Sink
 {
 public:
@@ -120,18 +95,6 @@ struct ConfigParametersNetworkSink
         std::nullopt,
         [](const std::unordered_map<std::string, std::string>& config) { return DescriptorConfig::tryGet(CHANNEL, config); }};
 
-    static inline const DescriptorConfig::ConfigParameter<size_t> BACKPRESSURE_UPPER_THRESHOLD{
-        "backpressure_upper_threshold",
-        1000,
-        [](const std::unordered_map<std::string, std::string>& config)
-        { return DescriptorConfig::tryGet(BACKPRESSURE_UPPER_THRESHOLD, config); }};
-
-    static inline const DescriptorConfig::ConfigParameter<size_t> BACKPRESSURE_LOWER_THRESHOLD{
-        "backpressure_lower_threshold",
-        200,
-        [](const std::unordered_map<std::string, std::string>& config)
-        { return DescriptorConfig::tryGet(BACKPRESSURE_LOWER_THRESHOLD, config); }};
-
     /// Per-channel sender queue size override. 0 means use the worker-level default.
     static inline const DescriptorConfig::ConfigParameter<size_t> SENDER_QUEUE_SIZE{
         "sender_queue_size",
@@ -146,14 +109,7 @@ struct ConfigParametersNetworkSink
 
     static inline std::unordered_map<std::string, DescriptorConfig::ConfigParameterContainer> parameterMap
         = DescriptorConfig::createConfigParameterContainerMap(
-            SinkDescriptor::parameterMap,
-            DATA_ENDPOINT,
-            CHANNEL,
-            BIND,
-            BACKPRESSURE_UPPER_THRESHOLD,
-            BACKPRESSURE_LOWER_THRESHOLD,
-            SENDER_QUEUE_SIZE,
-            MAX_PENDING_ACKS);
+            SinkDescriptor::parameterMap, DATA_ENDPOINT, CHANNEL, BIND, SENDER_QUEUE_SIZE, MAX_PENDING_ACKS);
 };
 
 /// NOLINTEND(cert-err58-cpp)
